+++ src/common/stop.c
@@ -19,6 +19,7 @@
 #include "events/filetailer.h"
 #include "system/linux/ipmi.h"
 #include "system/linux/nvml.h"
+#include "system/linux/dcgm.h"
 #include <signal.h>
 #include <string.h>
 #include <stdio.h>
@@ -59,6 +60,7 @@ void alligator_shutdown_after_loop(void)
 	glog(L_INFO, "Don't forget to start me again, otherwise the alligator will bite you :)\n");
 	ipmi_wait_idle();
 	nvml_wait_idle();
+	dcgm_wait_idle();
 	metric_dump(1);
 	tls_fs_free();
 	puppeteer_done();